+++ odoo/local-src/purchase_variant_fast_entry/__openerp__.py
@@ -1,47 +1,17 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Guewen Baconnier
-#    Copyright 2014 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# Copyright 2014-2016 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
+
 {'name': 'Purchase - variants fast entry',
- 'version': '1.2',
+ 'version': '9.1.0',
  'category': '',
- 'description': """
-Purchase - variants fast entry
-==============================
-
-Adds a button on the purchase orders.
-When used, it displays a wizard with a 'product template' field.
-Once the template is chosen, the wizard create one purchase line
-for each variant of this template.
-
-""",
  'author': 'Camptocamp',
  'maintainer': 'Camptocamp',
  'website': 'http://www.camptocamp.com/',
  'depends': ['purchase',
-             'product_variant_simple',
              ],
  'data': ['wizard/purchase_variant_fast_entry_view.xml',
-          'purchase_view.xml',
+          'views/purchase_views.xml',
           ],
- 'test': [],
- 'installable': False,
- 'auto_install': False,
- 'application': False,
+ 'installable': True,
  }