+++ HISTORY.rst
@@ -25,6 +25,7 @@ latest (unreleased)
 
 * Connector: transfer QoQa's payment id to account move lines'
   ``transaction_ref``
+* Migrate addon to create a purchase line for each variant of a template
 
 **Bugfixes**